Allow doggos to put on hats
This is a first draft version of the doggo hats patch by aoei.
Several folk have complained to K2 (Keith Simpson) that hatless
doggos are too vulnerable to nasty zombie brain eating and illness
attacks. Yes indeed! Poor doggos :-(

Thus the community have been demanding that K2 give doggos hats.
K2 has refused, adamantly, so I've taken things into my own hands...
Description of the patch: add two lines of code to include/mondata.h
that define a macro is_doggo() that returns true if mondata refers to
a little dog, dog, or large dog (kitties and horsies so far excluded).
Add two lines to m_dowear() in src/worn.c that allow us to skip the
check disallowing animals from putting on armour.

Issues with the current behaviour (as pointed out by aosdict):
having doggos put on their own hats from their inventory is unrealistic,
although perhaps a skilled doggo could put a hat on by picking it up
from the ground with its head, or (considering its inventory is its
mouth most probably), tossing it upward with a motion of the head, and
catching it on its crown. There, I have given a plausible description.

Having said all that, future versions of this patch may allow instead
the player to (a)pply a hat in the direction of a doggo (maybe even a
cat or horsie too) and put the helmet on that way.
Other possible additions may be to allow pet doggos to enter the
dungeon with a hat already equipped.

Signed-of-by: aoei / Joanna Janet Zaitseva-Doyle <jjadoyle@gmail.com>